relocate file
